SAP: make the update_count a u8.
It can only have three values 0,1,2, so storing a whole usize is useless.

+++ src/geometry/broad_phase_multi_sap.rs
@@ -340,7 +340,7 @@ struct SAPRegion {
     existing_proxies: BitVec,
     #[cfg_attr(feature = "serde-serialize", serde(skip))]
     to_insert: Vec<usize>, // Workspace
-    update_count: usize,
+    update_count: u8,
     proxy_count: usize,
 }
